+++ src/main/java/com/green/beadalyo/jhw/useraddr/model/UserAddrPostReq.java
@@ -5,6 +5,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class UserAddrPostReq {
@@ -17,7 +19,7 @@ public class UserAddrPostReq {
     @Schema(defaultValue = "상세주소")
     private String addr2;
     @Schema(defaultValue = "124.014")
-    private float addrCoorX;
+    private BigDecimal addrCoorX;
     @Schema(defaultValue = "36.746")
-    private float addrCoorY;
+    private BigDecimal addrCoorY;
 }
